requirements.txt
pygame
pyserial
kociemba
numpy
opencv-python
